# Optimize slices_from_chunks