Instead of taking Cartesian products of the starts and sizes, then
assembling the slices for every element of the n-d grid, first compute
the slices along each dimension and then take a Cartesian product.

Also make use of `cached_cumsum`, although that is somewhat incidental.

Before:
```
%timeit slices_from_chunks(((2,) * 1000, (3,) * 1000, (4,) * 10))
10.1 s ± 102 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

After:
```
%timeit slices_from_chunks(((2,) * 1000, (3,) * 1000, (4,) * 10))
629 ms ± 11.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

That's a 16x performance improvement. More practically, this knocked 20%
off the time for some of our code to construct a (large) graph.

- [ ] Tests added / passed
- [x] Passes `black dask` / `flake8 dask`