@bmerry 2020-02-13
I didn't add any tests because I'm trusting the existing tests to adequately test `slices_from_chunks`. Master has some failing tests at the moment, so I'm expecting the tests might fail on this branch too.