@mrocklin 2020-02-14
Yeah, not adding any tests for things like this sounds fine to me.  

Those speed benefits are impressive.  Thank you for tracking this down and fixing it.  Merging.