@rafa-guedes 2020-06-29
Some question related to this function optimised here. Reading metadata/coordinates from some large zarr archive (with nearly 5e6 objects) stored in GCS takes sometimes around 20 sec (from the same GCS region) with most of that time spent on 40 calls to `slices_from_chunks` (one for each data variable in this zarr archive). Would there be any way to try and circumvent these relatively costly calls? Reducing the number of chunks by increasing the chunk sizes has a large effect in reducing those times, but the original chunking was optimal for the way this data is usually read.