@jakirkham 2020-06-29
That seems to be a different concern than what this PR is changing. Could you please raise a new issue with an MRE of the problem you are seeing, @rafa-guedes?