@bmerry 2020-06-29
@rafa-guedes as @jakirkham suggests, a separate issue would be better. But that sounds like you're spending about 4µs per object, and I think you're going to struggle to do much better than that in Python, and is presumably small compared to the time to actually retrieve an object from GCS. If the issue is that you only want to actually use a small fraction of the data but are paying this cost for all the data, then you might be interested in #5918, although currently I haven't had time to do anything about it.