@rafa-guedes 2020-06-30
Thanks @jakirkham, @bmerry, I'll open a new issue on this